refactor(backend): Move poll service and controller
