+++ apps/backend/src/poll/poll.service.ts
@@ -23,10 +23,10 @@ import {Injectable, Logger, NotFoundException, OnModuleInit, UnprocessableEntity
 import {InjectModel} from '@nestjs/mongoose';
 import {Document, FilterQuery, Model, Types} from 'mongoose';
 
-import {environment} from '../../environment';
-import {renderDate} from '../../mail/helpers';
-import {MailService} from '../../mail/mail/mail.service';
-import {PushService} from '../../push/push.service';
+import {environment} from '../environment';
+import {renderDate} from '../mail/helpers';
+import {MailService} from '../mail/mail/mail.service';
+import {PushService} from '../push/push.service';
 
 @Injectable()
 export class PollService implements OnModuleInit {
